requirements.txt
python-dotenv
telethon
google-genai
fpdf2
aiogram
httpx